chore(extensions): Refactor extensions error logic
