+++ src/extensions/extensions.ts
@@ -17,7 +17,8 @@
 
 import { App } from '../app';
 import { SettableProcessingState } from './extensions-api';
-import { ExtensionsApiClient, FirebaseExtensionsError } from './extensions-api-client-internal';
+import { ExtensionsApiClient } from './extensions-api-client-internal';
+import { FirebaseExtensionsError } from './error';
 import * as validator from '../utils/validator';
 
 /**
